Simplify required callbacks for Backtrex behaviour.
It's much more natural to describe the valid values for cells in a
Sudoku solution as `1..9` than the convoluted way it was done with
`c:initial_value/2` and `c:next_value/3`.

I also added back logging, but by default all logs below the `:warn`
level are purged at compile time. The tests run are pretty fast this
way, but are significantly slower when `:info` and `:debug` are on.