Flame graphs using :eflame
The example puzzle is the same as the one in the Sudoku solver test, except with
several more cells filled in. Using the original causes :eflame to time out
while creating the final report (see
https://github.com/proger/eflame/issues/13).

Ensure rebar is installed (the original, not rebar3), and run the following:

    $ MIX_ENV=prod mix do clean, deps.get, deps.compile, compile
    $ MIX_ENV=prod iex -S mix run -e "Backtrex.Profiler.eflame; exit(:normal)"
    $ ./deps/eflame/stack_to_flame.sh < stacks.out > flame-prod.svg

Credit goes to http://learningelixir.joekain.com/profiling-elixir-2/ for getting
me started.